requirements.txt
Flask
gunicorn
Pillow
python-dotenv
opencv-python-headless
numpy
rembg[cpu]